+++ backend/src/role/index.ts
@@ -7,7 +7,7 @@ export {
   ROLE_CHANGED_CHANNEL,
 } from './RoleService'
 export { DEFAULT_ROLES } from './defaults'
-export { ensureUserRoleEdges } from './userRoleEdges'
+export { ensureUserRoleEdges, seedDefaultRoleNodes } from './userRoleEdges'
 export { effectiveRoleName } from './effectiveRoleNames'
 export { OWNER_ROLE, ADMIN_ROLE, MODERATOR_ROLE, USER_ROLE } from './types'
 export type { RoleDefinition, RoleChangeEvent, RolePubSub } from './types'
